fix: move the object to another module
